+++ src/operators/tensor/implementations/tensor_fp16x16wide.cairo
@@ -313,7 +313,7 @@ impl FP16x16WTensor of TensorTrait<FP16x16W> {
     }
 
     fn gather(
-        self: @Tensor<FP16x16W>, indices: Tensor<usize>, axis: Option<usize>
+        self: @Tensor<FP16x16W>, indices: Tensor<i32>, axis: Option<i32>
     ) -> Tensor<FP16x16W> {
         math::gather::gather(self, indices, axis)
     }
